feat: rename job event presentation module and add tests for event formatting
