+++ apps/sva-studio-react/src/routes/monitoring/-job-event-presentation.test.ts
@@ -7,7 +7,7 @@ import {
   formatMonitoringJobEventTitle,
   resolveMonitoringJobEventIsTerminal,
   resolveMonitoringJobEventTone,
-} from './job-event-presentation';
+} from './-job-event-presentation';
 
 const baseEvent: StudioJobEventRecord = {
   id: 'event-1',
